requirements.txt
pillow
origamibot
